requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
python-docx==1.1.0
img2pdf==0.5.1
PyMuPDF==1.23.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
qrcode>=7.4.2
Pillow>=10.1.0
requests>=2.31.0
python-dotenv>=1.0.0
razorpay>=1.4.1
